# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Orchis Foundry API",
    "version": "0.7.0",
    "description": "Self-hosted code platform: git hosting, pull requests, per-user BYO-model supply-chain scanning, and LLM-friendly context endpoints. Authenticate with a personal access token: `Authorization: Bearer orc_pat_<...>`."
  },
  "servers": [{ "url": "/v1" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "orc_pat" }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/pulls": { "get": { "summary": "Pull requests across your repos", "parameters": [{ "name": "filter", "in": "query", "schema": { "type": "string", "enum": ["review-requested", "yours", "mentioned", "all"] } }], "responses": { "200": { "description": "[PullSummary]" } } } },
    "/repos/{org}/{name}/pulls": {
      "get": { "summary": "List PRs in a repo", "parameters": [{ "name": "state", "in": "query", "schema": { "type": "string", "enum": ["open", "closed", "merged", "all"] } }], "responses": { "200": { "description": "[PullSummary]" } } },
      "post": { "summary": "Open a PR", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "title": { "type": "string" }, "head": { "type": "string" }, "base": { "type": "string" }, "body": { "type": "string" } }, "required": ["title", "head"] } } } }, "responses": { "200": { "description": "PullSummary" } } }
    },
    "/repos/{org}/{name}/pulls/{num}": {
      "get": { "summary": "PR detail", "responses": { "200": { "description": "PullSummary" } } },
      "patch": { "summary": "Edit PR title/body", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "title": { "type": "string" }, "body": { "type": "string" } } } } } }, "responses": { "200": { "description": "PullSummary" } } }
    },
...
  }
}